+++ user_code/protocol/Referee.h
@@ -328,7 +328,7 @@ public:
     //42mm枪口热量上限, 42mm枪口实时热量
     void get_shooter_id1_42mm_cooling_limit_and_heat(uint16_t *id1_42mm_cooling_limit, uint16_t *id1_42mm_cooling_heat);
     //42mm枪口枪口射速上限,42mm实时射速
-    void get_shooter_id1_42mm_speed_limit_and_bullet_speed(uint16_t *id1_42mm_speed_limit, uint16_t *bullet_speed);
+    void get_shooter_id1_42mm_speed_limit_and_bullet_speed(uint16_t *id1_42mm_speed_limit, fp32 *bullet_speed);
     //42mm枪口热量冷却
     void get_shooter_id1_42mm_cooling_rate(uint16_t *id1_42mm_cooling_rate);
     //当前血量
